backend/app/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers
from rest_framework_simplejwt.tokens import RefreshToken

class UserSerializer(serializers.ModelSerializer):
    class Meta: 
        model = User
        fields = (
            "id",
            "user",
            "token"
        )

